requirements.txt
streamlit
matplotlib
plotly
spacy
textblob
wordcloud
newspaper3k